refactor(components): add code style conventions and apply to existing components
Documents code style rules in CLAUDE.md (TSDoc comments, component folder
structure, direct React imports, default exports for components, named exports
for hooks/constants, function declarations, accessibility, whitespace).

Applies those rules across all existing custom components:
- Move flat component files into named subfolders (e.g. EventCard/EventCard.tsx)
- Replace `import * as React` with direct named imports
- Convert named component exports to default exports
- Add whitespace above return statements and around if blocks
- Fix accessibility: aria-label on icon-only close button, htmlFor/id on
  zip code label/input, aria-label on radius slider
- Add TSDoc comments for non-obvious props with defaults
