+++ src/components/ui/sheet.tsx
@@ -1,9 +1,9 @@
 import * as React from 'react';
 import { Dialog as SheetPrimitive } from 'radix-ui';
 
+import { XIcon } from '@phosphor-icons/react';
 import { cn } from '@/lib/utils';
 import { Button } from '@/components/ui/button';
-import { XIcon } from '@phosphor-icons/react';
 
 function Sheet({ ...props }: React.ComponentProps<typeof SheetPrimitive.Root>) {
   return <SheetPrimitive.Root data-slot="sheet" {...props} />;
