+++ src/components/ui/toggle-group.tsx
@@ -1,8 +1,8 @@
 'use client';
 
 import * as React from 'react';
-import { type VariantProps } from 'class-variance-authority';
 import { ToggleGroup as ToggleGroupPrimitive } from 'radix-ui';
+import type { VariantProps } from 'class-variance-authority';
 
 import { cn } from '@/lib/utils';
 import { toggleVariants } from '@/components/ui/toggle';
