+++ src/components/ui/toggle.tsx
@@ -1,6 +1,7 @@
 import * as React from 'react';
-import { cva, type VariantProps } from 'class-variance-authority';
+import { cva } from 'class-variance-authority';
 import { Toggle as TogglePrimitive } from 'radix-ui';
+import type { VariantProps } from 'class-variance-authority';
 
 import { cn } from '@/lib/utils';
